hivemind/language-modeling/netmind/trainer_customer/requirements.txt
transformers>=4.6.0
datasets==1.18.3
torch_optimizer>=0.1.0
wandb>=0.10.26
sentencepiece
requests
nltk>=3.6.2
boto3
retry
setuptools==59.5.0